Move agent RBACs to the component
